reorganized auth into a Blueprint
