+++ app/auth/routes.py
@@ -1,0 +1,66 @@
+import pam
+from flask import flash, redirect, render_template, request, url_for
+from flask_login import current_user, login_user, logout_user
+from werkzeug.urls import url_parse
+
+from app import db
+from app.auth import bp
+from app.auth.forms import LoginForm
+from app.models import User
+
+FLASH_CLS = {
+    "error": "alert alert-danger",
+    "success": "alert alert-success",
+}
+
+
+@bp.route("/login", methods=["GET", "POST"])
+def login():
+    if current_user.is_authenticated:
+        return redirect(url_for("home.index"))
+    form = LoginForm(request.form)
+    if request.method == "POST":
+        if form.validate():
+            user = User.query.filter_by(username=form.username.data).first()
+            if user is None:
+                if pam.authenticate(form.username.data, form.password.data):
+                    try:
+                        user = User(username=form.username.data)
+                        user.set_password_hash(form.password.data)
+                        db.session.add(user)
+                        db.session.commit()
+                    except Exception as e:
+                        db.session.rollback()
+                        flash(
+                            message=(
+                                "Error adding user to database. "
+                                "Contact an admin for assistance."
+                            ),
+                            category=FLASH_CLS["error"],
+                        )
+                        return redirect(url_for("auth.login"))
+                else:
+                    flash(
+                        message="Invalid username or password",
+                        category=FLASH_CLS["error"],
+                    )
+                    return redirect(url_for("auth.login"))
+            if not user.check_password(form.password.data):
+                flash(
+                    message="Invalid username or password", category=FLASH_CLS["error"]
+                )
+                return redirect(url_for("auth.login"))
+            login_user(user, remember=form.remember_me.data)
+            next_page = request.args.get("next")
+            if not next_page or url_parse(next_page).netloc != "":
+                next_page = url_for("home.index")
+            return redirect(next_page)
+        else:
+            flash(message="All fields required.", category=FLASH_CLS["error"])
+    return render_template("auth/login.html", form=form)
+
+
+@bp.route("/logout")
+def logout():
+    logout_user()
+    return redirect(url_for("auth.login"))
