--- app/extensions.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python3
-
-from flask_caching import Cache
-from flask_login import LoginManager
-from flask_migrate import Migrate
-from flask_sqlalchemy import SQLAlchemy
-
-db = SQLAlchemy()
-migrate = Migrate()
-login = LoginManager()
-cache = Cache()
