+++ app/models.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from flask_login import UserMixin
 from notebook.auth.security import passwd, passwd_check
 
-from .extensions import db, login
+from app import db, login
 
 
 class ActivityLog(db.Model):
