+++ app/views/home.py
@@ -1,15 +1,13 @@
 from datetime import datetime
 from urllib.parse import urlparse
 
-import pam
 from flask import Blueprint, flash, redirect, render_template, request, url_for
-from flask_login import current_user, login_required, login_user, logout_user
+from flask_login import current_user, login_required
 
 from .. import launch
-from ..extensions import cache, db
-from ..forms import LoginForm
-from ..helpers import data_query
-from ..models import ActivityLog, User
+from app import db
+from app.helpers import data_query
+from app.models import ActivityLog
 
 FLASH_CLS = {
     "error": "alert alert-danger",
@@ -144,52 +142,6 @@ def kill_session():
     return redirect(url_for("home.index"))
 
 
-@home.route("/login", methods=["GET", "POST"])
-def login():
-    if current_user.is_authenticated:
-        return redirect(url_for("home.index"))
-    form = LoginForm(request.form)
-    if request.method == "POST":
-        if form.validate():
-            user = User.query.filter_by(username=form.username.data).first()
-            if user is None:
-                if pam.authenticate(form.username.data, form.password.data):
-                    try:
-                        user = User(username=form.username.data)
-                        user.set_password_hash(form.password.data)
-                        db.session.add(user)
-                        db.session.commit()
-                    except Exception as e:
-                        db.session.rollback()
-                        flash(
-                            message=(
-                                "Error adding user to database. "
-                                "Contact an admin for assistance."
-                            ),
-                            category=FLASH_CLS["error"],
-                        )
-                        return redirect(url_for("home.login"))
-                else:
-                    flash(
-                        message="Invalid username or password",
-                        category=FLASH_CLS["error"],
-                    )
-                    return redirect(url_for("home.login"))
-            if not user.check_password(form.password.data):
-                flash(
-                    message="Invalid username or password", category=FLASH_CLS["error"]
-                )
-                return redirect(url_for("home.login"))
-            login_user(user, remember=form.remember_me.data)
-            next_page = request.args.get("next")
-            if not next_page or urlparse(next_page).netloc != "":
-                next_page = url_for("home.index")
-            return redirect(next_page)
-        else:
-            flash(message="All fields required.", category=FLASH_CLS["error"])
-    return render_template("login.html", form=form)
-
-
 @home.route("/dashboard", methods=["GET"])
 @login_required
 def dashboard():
@@ -202,9 +154,3 @@ def data():
     start_date = request.args.get("start_date")
     end_date = request.args.get("end_date")
     return data_query(start_date, end_date)
-
-
-@home.route("/logout")
-def logout():
-    logout_user()
-    return redirect(url_for("home.login"))
